backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "google-genai>=1.43.0",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.35.3",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pymupdf>=1.26.5",
    "vosk>=0.3.45",
]
